+++ NEWS.md
@@ -32,6 +32,8 @@
 - A better 'List sort' wizard. (Mentioned in the discussion forum).
 - Fix the Help->Example and Help->Apropos menus.
 - Fix the Matrix->Row and columns operations->Remove rows or columns menu.
+- Fix statistics sidebar
+  (wxbarsplot, wxboxplot, wxpiechart, mean, median, std were not working)
 
 # 26.01.0